win-well-guide-main/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
# FastAPI and LangChain dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain>=1.2.0
langchain-core>=1.2.2
langchain-community>=0.3.0
langchain-ollama>=1.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic>=2.7.4
